change the way to manager the default product

+++ src/components/ProductList.tsx
@@ -13,19 +13,20 @@ function ProductList({ productList, deleteProductFunction, checkProductFunction 
     <>
       <ul className='product-list'>
         {
-            productList.map(product => {
+            productList.length === 1 && productList[0].id === "DefaultProduct"
+            ? <li style={{listStyle: "none"}}>No products yet</li>
+            : productList.map(product => {
                 const productCheckClassName = product.check ? "product--green" : "product--red"
                 const productClassName = `product ${productCheckClassName}`
+                if(product.id === "DefaultProduct") return;
                 return (
                     <li key={product.id} className={productClassName}>
                         { product.productName }
                         <button onClick={() => checkProductFunction(product.id)}>Check</button>
                         <button onClick={() => deleteProductFunction(product.id)}>Delete</button>
-                    </li>
+                     </li>
                 )
-            }
-               
-            )
+            })
         }
       </ul>
     </>
